+++ packages/mobile/src/components/Loading/Loading.stories.tsx
@@ -1,6 +1,11 @@
-import { storiesOf } from '@storybook/react-native'
 import React from 'react'
+import { storiesOf } from '@storybook/react-native'
 
 import { Loading } from './Loading.component'
 
-storiesOf('Loading', module).add('Default', () => <Loading />)
+storiesOf('Loading', module).add('Default', () => (
+  <Loading
+    title={'Creating community “Disco-fever”'}
+    caption={'Additional info if needed can go here otherwise this is hidden.'}
+  />
+))